fix uninitialized ASN.1 objects use
The underlying pyasn1 library has a notion of `schema` and
`value` object. The `schema` object can't be used with
regular Python operators as values because `schema` does
not carry any value, it's rather a type.

This fix turns a couple of `NoSuchInstance` schema
objects into proper `value` objects to retain compatibility
with the latest pysnmp which removed the default initializer
of the `NoSuchInstance` object.

Change-Id: Id044a0e437724bf1f3581e9187e2a1e6c336b1bc